updated the getallcustomers method to use long instead of int for the phonenumber

+++ DL/CustomerRepository.cs
@@ -41,7 +41,7 @@ namespace DL
                         State = rest.State,
                         Address = rest.Address,
                         Email = rest.Email,
-                        PhoneNumber = (int)rest.PhoneNumber
+                        PhoneNumber = (long)rest.PhoneNumber
                     }
             ).ToList();
         }
